Add new api features and refactory the code
- Add the feature: phaseTypes
- Add the feature: activityTypes
- Move middleware to server/api/middleware.js
- Extract utils functions from middleware and add them to
server/api/utils.js
